pyproject.toml
[project]
name = "openai-api-experiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
     "python-dotenv",
     "openai",
     "pandas"
]
